+++ vertx-client/src/main/java/com/cloudentity/tools/vertx/http/builder/RequestCtxBuilder.java
@@ -6,6 +6,7 @@ import io.vertx.core.Handler;
 import io.vertx.core.buffer.Buffer;
 import io.vertx.core.http.HttpClientResponse;
 import io.vertx.core.http.HttpMethod;
+import io.vertx.core.streams.Pipe;
 
 public interface RequestCtxBuilder extends CallValuesBuilder<RequestCtxBuilder> {
   RequestCtxBuilder method(HttpMethod m);
@@ -14,24 +15,30 @@ public interface RequestCtxBuilder extends CallValuesBuilder<RequestCtxBuilder> 
   Future<HttpClientResponse> end();
   Future<HttpClientResponse> end(Buffer b);
   Future<HttpClientResponse> end(String b);
+  Future<HttpClientResponse> end(Pipe<Buffer> s);
 
   Future<HttpClientResponse> end(Handler<Buffer> bodyHandler);
   Future<HttpClientResponse> end(Buffer b, Handler<Buffer> bodyHandler);
   Future<HttpClientResponse> end(String b, Handler<Buffer> bodyHandler);
+  Future<HttpClientResponse> end(Pipe<Buffer> s, Handler<Buffer> bodyHandler);
 
   Future<SmartHttpResponse> endWithBody();
   Future<SmartHttpResponse> endWithBody(Buffer b);
   Future<SmartHttpResponse> endWithBody(String b);
+  Future<SmartHttpResponse> endWithBody(Pipe<Buffer> s);
 
   Future<HttpClientResponse> end(TracingContext tracingContext);
   Future<HttpClientResponse> end(TracingContext tracingContext, Buffer b);
   Future<HttpClientResponse> end(TracingContext tracingContext, String b);
+  Future<HttpClientResponse> end(TracingContext tracingContext, Pipe<Buffer> s);
 
   Future<HttpClientResponse> end(TracingContext tracingContext, Handler<Buffer> bodyHandler);
   Future<HttpClientResponse> end(TracingContext tracingContext, Buffer b, Handler<Buffer> bodyHandler);
   Future<HttpClientResponse> end(TracingContext tracingContext, String b, Handler<Buffer> bodyHandler);
+  Future<HttpClientResponse> end(TracingContext tracingContext, Pipe<Buffer> s, Handler<Buffer> bodyHandler);
 
   Future<SmartHttpResponse> endWithBody(TracingContext tracingContext);
   Future<SmartHttpResponse> endWithBody(TracingContext tracingContext, Buffer b);
   Future<SmartHttpResponse> endWithBody(TracingContext tracingContext, String b);
+  Future<SmartHttpResponse> endWithBody(TracingContext tracingContext, Pipe<Buffer> s);
 }